TODO cleanup pass
