+++ hydra/_internal/grammar/functions.py
@@ -1,12 +1,11 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
 import inspect
 from dataclasses import dataclass, field
-from inspect import Signature
-from typing import Any, Callable, Dict, List, Optional, Type, Union
+from typing import Any, Callable, Dict, List
 
-# TODO: move function related stuff to internal
-from omegaconf._utils import is_dict_annotation, is_list_annotation, type_str
+from omegaconf._utils import type_str
 
+from hydra._internal.grammar.utils import is_type_matching
 from hydra.core.override_parser.types import QuotedString
 from hydra.errors import HydraException
 
@@ -18,41 +17,9 @@ class FunctionCall:
     kwargs: Dict[str, Any]
 
 
-def get_list_element_type(ref_type: Optional[Type[Any]]) -> Optional[Type[Any]]:
-    args = getattr(ref_type, "__args__", None)
-    if ref_type is not List and args is not None and args[0] is not Any:
-        element_type = args[0]
-    else:
-        element_type = None
-    assert element_type is None or issubclass(element_type, type)
-    return element_type  # type: ignore
-
-
-# TODO: move to some utils
-def is_type_matching(value: Any, type_: Any) -> bool:
-    # Union
-    if hasattr(type_, "__origin__") and type_.__origin__ is Union:
-        types = list(type_.__args__)
-        for idx, t in enumerate(types):
-            # for now treat any Dict[X,Y] as dict and any List[X] as list, ignoring element types
-            if is_dict_annotation(t):
-                t = dict
-            elif is_list_annotation(t):
-                t = list
-            types[idx] = t
-        return isinstance(value, tuple(types))
-    else:
-        primitives = (int, float, bool, str)
-        if type_ in primitives:
-            return type(value) is type_
-        if type_ in (Any, inspect.Signature.empty):
-            return True
-        return isinstance(value, type_)
-
-
 @dataclass
 class Functions:
-    definitions: Dict[str, Signature] = field(default_factory=dict)
+    definitions: Dict[str, inspect.Signature] = field(default_factory=dict)
     functions: Dict[str, Callable[..., Any]] = field(default_factory=dict)
 
     def register(self, name: str, func: Callable[..., Any]) -> None:
@@ -105,6 +72,4 @@ class Functions:
                         f" {type_str(type(value))} is incompatible with {type_str(expected_type)}"
                     )
 
-        bound_args = bound.args
-        bound_kwargs = bound.kwargs
-        return self.functions[func.name](*bound_args, **bound_kwargs)
+        return self.functions[func.name](*bound.args, **bound.kwargs)
